requirements.txt
pandas>=2.0
mysql-connector-python>=9.0
pyyaml>=6.0
python-dateutil>=2.9
python-dotenv>=1.0.0
